+++ tests/test_assays.py
@@ -525,6 +525,41 @@ def test_olink():
         validator.validate(obj)
 
 
+def test_microbiome():
+
+    # create the ngs object
+    ngs_obj = {"sequencer_platform": "Illumina - NovaSeq 6000"}
+    obj = {**ASSAY_CORE, **ngs_obj}  # merge two dictionaries
+
+    fastq = ARTIFACT_OBJ.copy()
+    fastq["data_format"] = "FASTQ.GZ"
+    tsv = ARTIFACT_OBJ.copy()
+    tsv["data_format"] = "TSV"
+
+    # add custom entry
+    obj["enrichment_method"] = "Ribo minus"
+    obj["enrichment_vendor_kit"] = "Agilent"
+    obj["batch_id"] = "batch1"
+    obj["forward_fastq"] = fastq
+    obj["forward_index"] = fastq
+    obj["reverse_fastq"] = fastq
+    obj["reverse_index"] = fastq
+    obj["otu_table"] = tsv
+
+    # create the microbiome object
+    record = {
+        "library_yield_ng": 666,
+        "cimac_id": "CTTTPPPSA.00",
+    }
+
+    # add a demo record.
+    obj["records"] = [record]
+
+    # create validator assert schemas are valid.
+    validator = _fetch_validator("microbiome")
+    validator.validate(obj)
+
+
 def test_clinicaldata():
 
     # create validator